implemented health pickups & fixed some issues

+++ Assets/Scripts/Gameplay/ScrolledObjects/IScrolledObjectLogic.cs
@@ -8,10 +8,10 @@ namespace Gameplay.ScrolledObjects
 
         public void ScrolledObjectFixedUpdate(ScrolledObjectView view);
 
-        public void OnHitByWeapon(ScrolledObjectView view, float damage);
+        public void OnHitByWeapon(ScrolledObjectView view, int damage);
 
         public void OnHitByPlayer(ScrolledObjectView view, Action<int> hpAction, Action<int> xpAction);
-
+        
         public void OnActivate(int value);
 
         public void OnDeactivate();
